StaleElementException fix (hopefully)

+++ TwitterListener/Listener.cs
@@ -12,7 +12,7 @@ namespace TwitterListener
             "/div[2]/div/div/article/div/div/div/div[2]/div[2]/div[1]/div/div/div[1]"; // This is the XPath of second relative Tweet in the feed
         private const string relativeThird = "/html/body/div[1]/div/div/div[2]/main/div/div/div/div[1]/div/div[2]/div/div/section/div/div" +
             "/div[3]/div/div/article/div/div/div/div[2]/div[2]/div[1]/div/div/div[1]"; // This is the XPath of third relative Tweet in the feed
-        // Both are valid for now ofc
+        // All are valid for now ofc
 
         public static IWebElement GetNewestTweetElement(WebDriver driver, ref bool exitState)
         {
@@ -38,11 +38,14 @@ namespace TwitterListener
                     }
                     return driver.FindElement(By.XPath(relativeFirst)); // Topmost re/Tweet is good to go
                 }
-                catch (NoSuchElementException) // If for some reason it can't find the element, it will throw this
+                catch (Exception ex) // If for some reason it can't find the element, it will throw this
                 {
-                    // What if we just deleted cookies and rerolled?
-                    driver.Manage().Cookies.DeleteAllCookies();
-                    WebdriverHandler.RefreshPageWithExceptionHandling(driver);
+                    if (ex is NoSuchElementException || ex is StaleElementReferenceException)
+                    {
+                        // What if we just deleted cookies and rerolled?
+                        driver.Manage().Cookies.DeleteAllCookies();
+                        WebdriverHandler.RefreshPageWithExceptionHandling(driver);
+                    }
                 }
             }
             return null;
